adds more image, fix mobile and desktop view
there was a bug that would refresh the page each time the animation ends, this is due to setting absolute on images that would require another refresh of the page for some reasons. the only place were absolute is used is for text based content, as it doesn't trigger the browser to refresh. basic implementation is now completed, more styling necessary but minimal implementation has been archived